fix(next-pwa): fixed build crashing when swcMinify is set to false
