+++ .changeset/eighty-bags-beg.md
@@ -1,0 +1,5 @@
+---
+"@ducanh2912/next-pwa": patch
+---
+
+fix(next-pwa): fixed build crashing when swcMinify is set to false
