+++ docs/content/next-pwa/configuring.mdx
@@ -79,6 +79,8 @@ export default withPWA({
 
 - `extendDefaultRuntimeCaching` — Extend the default `runtimeCaching` array. Only effective when `runtimeCaching` is specified.
 
+- `swcMinify` — Use [`swc`](https://swc.rs) to minify the custom worker, fallback worker,...
+
 ### Experimental
 
 - `aggressiveFrontEndNavCaching` — Cache every `<link rel="stylesheet" />` and `<script />` on frontend navigation. Requires `cacheOnFrontEndNav` to be enabled.
